% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcjaccard.R
\name{calcjaccard}
\alias{calcjaccard}
\title{Calculate Jaccard Similarity for a Pair of Topics}
\usage{
calcjaccard(topic1, topic2)
}
\arguments{
\item{topic1}{a vector with terms most representative of a topic.}

\item{topic2}{a vector with terms most representative of another topic.}
}
\value{
a value of Jaccard similarity for a pair of topics (numerical).
}
\description{
A primary internal function for calculating Jaccard similarity for a given
pair of topics.
}
\examples{
topicA <- c("tax", "money", "million", "financi", "govern", "spend", "billion")
topicB <- c("iraq", "war", "militari", "iraqi", "troop", "forc", "will")

calcjaccard(topicA, topicB)
}
