% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_solutions.R
\name{compare_solutions}
\alias{compare_solutions}
\title{Pair-Wisely Compare Topic Solutions with Jaccard}
\usage{
compare_solutions(models, depth = 100)
}
\arguments{
\item{models}{a list object with STM models. Must be at least 2 models.}

\item{depth}{the number of top terms to use for comparison. Default is top
100 terms.}
}
\value{
a tibble with model-to-model topic-to-topic Jaccard similarity.
}
\description{
Pair-Wisely Compare Topic Solutions with Jaccard
}
\examples{
library(stm)
library(dplyr)

modA <- stm(poliblog5k.docs, 
            poliblog5k.voc, K=25,
            prevalence=~rating, 
            data=poliblog5k.meta,
            max.em.its=2, 
            init.type="Random",
            seed = 9934) 
           
modB <- stm(poliblog5k.docs, 
            poliblog5k.voc, K=25,
            prevalence=~rating, 
            data=poliblog5k.meta,
            max.em.its=2, 
            init.type="Random",
            seed = 9576) 
           
compare_solutions(list(modA, modB), depth=100) |> 
arrange(desc(jaccard)) |> 
head()
}
