% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_solutions.R
\name{filter_topics}
\alias{filter_topics}
\title{Summarize solution comparisons}
\usage{
filter_topics(res, sim_threshold, n_sols)
}
\arguments{
\item{res}{output from \code{compare_solutions()}}

\item{sim_threshold}{level of Jaccard similarity for filtering stable topics.}

\item{n_sols}{the number of fits to decide if a topic is stable.}
}
\value{
tibble with counts of topic persistence across fits.
}
\description{
Summarize solution comparisons
}
