% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_terms.R
\name{top_terms}
\alias{top_terms}
\title{Get Top Terms for a Topic}
\usage{
top_terms(mod, topic_id, n_terms)
}
\arguments{
\item{mod}{an object with STM or LDA model.}

\item{topic_id}{topic ID (numeric).}

\item{n_terms}{the number of terms to return.}
}
\value{
A tibble with a ranked list of terms.
}
\description{
Returns top-N terms for a topic ranked by probability.
}
\examples{
library(stm)

mod <- stm(poliblog5k.docs, 
           poliblog5k.voc, K=25,
           prevalence=~rating, 
           data=poliblog5k.meta,
           max.em.its=2, 
           init.type="Spectral") 

top_terms(mod, 1, 100)
}
