% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_solutions.R
\name{viz_comparisons}
\alias{viz_comparisons}
\title{Visualize topic similarities with a network}
\usage{
viz_comparisons(res, sim_threshold)
}
\arguments{
\item{res}{output of \code{compare_solutions()}}

\item{sim_threshold}{Jaccard similarity threshold for filtering.}
}
\value{
ggraph plot.
}
\description{
Visualize topic similarities with a network
}
